requirements.txt
pandas~=2.2.3
fitz~=0.0.1.dev2
docx~=0.2.4
python-pptx~=1.0.2